Project persistence repo - get map tests written and done

+++ Java/src/main/java/com/bae/persistence/repository/PoseDatabaseRepository.java
@@ -32,7 +32,7 @@ public class PoseDatabaseRepository implements PoseRepository {
 		return json.getJSONForObject(accounts);
 	}
 
-	public String getAPose(long poseId) {
+	public String getAPose(int poseId) {
 		return json.getJSONForObject(manager.find(Pose.class, poseId));
 	}
