+++ Java/src/main/java/com/bae/persistence/repository/PoseRepository.java
@@ -2,6 +2,7 @@ package com.bae.persistence.repository;
 
 public interface PoseRepository {
 	String getAllPoses();
-	String getAPose(long poseID);
+
+	String getAPose(int poseID);
 
 }
